Buffer optimizations, fixed approximation fog, improved reflections scaling, cleaned code

+++ shaders/internalSettings.glsl
@@ -18,15 +18,15 @@ const int shadowcolor1Format = RGBA8;
 
 const int colortex0Format  = RGBA16F;   // Clouds
 const int colortex1Format  = RGBA32UI;  // Gbuffers Data
-const int colortex2Format  = RGBA16F;   // Geometric normals, POM selfshadowing and reflections
-const int colortex3Format  = RGBA16F;   // Shadowmap, Bloom
-const int colortex4Format  = RGBA32F;   // Main Buffer 0
+const int colortex2Format  = RGBA16F;   // Reflections
+const int colortex3Format  = RGBA16F;   // Geometric normals, POM selfshadowing, Shadowmap, Bloom
+const int colortex4Format  = RGBA16F;   // Main Buffer 0
 const int colortex5Format  = RGBA32F;   // Main Buffer 1
 const int colortex6Format  = RGB32F;    // Direct & Indirect Illuminances
 const int colortex8Format  = RGBA16F;   // History Buffer
-const int colortex9Format  = RGBA32F;   // Direct PT
-const int colortex10Format = RGBA32F;   // Indirect PT
-const int colortex11Format = RGBA32F;   // Moments
+const int colortex9Format  = RGBA16F;   // Direct PT
+const int colortex10Format = RGBA16F;   // Indirect PT
+const int colortex11Format = RGBA16F;   // Moments
 const int colortex12Format = RGB16F;    // Sky
 const int colortex13Format = RGBA16F;   // Translucents
